Add `DevicesBuf` helper for building device descriptors.
The `sane_get_devices()` function returns a pointer to a
`NULL`-terminated array of `SANE_Device*`, and that pointer must remain
valid until the next call to `sane_get_devices()` or `sane_exit()`.

A `DevicesBuf` can be used to track ownership of the `Device` structures
and their allocated string fields.

GitOrigin-RevId: 7259347f17745e38e0eba17769cdd73d72427767