Add `DevicesBuf` helper for building device descriptors.
The `sane_get_devices()` function returns a pointer to a
`NULL`-terminated array of `SANE_Device*`, and that pointer must remain
valid until the next call to `sane_get_devices()` or `sane_exit()`.

A `DevicesBuf` can be used to track ownership of the `Device` structures
and their allocated string fields.

GitOrigin-RevId: 57097e8237afa29641fa059987191db3085872b2